+++ nix/lib/default.nix
@@ -38,6 +38,8 @@ in
       listOfOutputType ? conanFlakeLib.types.listOfOutputType lib,
       anyOutput ? conanFlakeLib.types.anyOutput lib,
       mergeSelected ? conanFlakeLib.mergeSelected lib,
+      pnameFromStdenvCc ? conanFlakeLib.pnameFromStdenvCc lib,
+      versionFromStdenvCc ? conanFlakeLib.versionFromStdenvCc lib,
     }:
     {
       inherit (conanFlakeLib) contains;
@@ -52,6 +54,8 @@ in
         listOfOutputType
         anyOutput
         mergeSelected
+        pnameFromStdenvCc
+        versionFromStdenvCc
         ;
     };
 
@@ -160,4 +164,28 @@ in
   mergeSelected =
     lib: f: select: attrs:
     lib.mkMerge (lib.mapAttrsToList select (lib.filterAttrs f attrs));
+
+  pnameFromStdenvCc =
+    lib: stdenv:
+    let
+      cc = stdenv.cc;
+      name = cc.meta.name;
+      components = lib.splitString "-" name;
+      length = builtins.length components;
+      pnameLength = if length >= 1 then length - 1 else 0;
+      pnameComponents = lib.lists.take pnameLength components;
+    in
+    builtins.concatStringsSep "" pnameComponents;
+
+  versionFromStdenvCc =
+    lib: stdenv:
+    let
+      cc = stdenv.cc;
+      name = cc.meta.name;
+      components = lib.splitString "-" name;
+      length = builtins.length components;
+      pnameLength = if length >= 1 then length - 1 else 0;
+      versionLength = length - pnameLength;
+    in
+    if versionLength > 0 then builtins.elemAt components (length - 1) else "";
 }