+++ test/flake-parts-no-defaults/flake.nix
@@ -52,7 +52,9 @@
             inherit configLocal conanHome profiles;
 
             settings.compiler = {
-              "${cfg.stdenv.cc.cc.pname}".version.__assign = [ cfg.stdenv.cc.version ];
+              "${inputs.conan-flake.lib.pnameFromStdenvCc lib cfg.stdenv}".version.__assign = [
+                (inputs.conan-flake.lib.versionFromStdenvCc lib cfg.stdenv)
+              ];
             };
 
             offline = true;
@@ -98,7 +100,7 @@
                       echo "Checking local setup..."
 
                       cat ${escapeShellArg cfg.configLocal}"/settings_user.yml" \
-                        | grep -F ${escapeShellArg cfg.stdenv.cc.version}
+                        | grep -F ${escapeShellArg (inputs.conan-flake.lib.pnameFromStdenvCc lib cfg.stdenv)}
 
                       cat ${escapeShellArg cfg.configLocal}"/profiles/default" \
                         | grep -F "build_type="${escapeShellArg cfg.final.profiles.settings._.build_type}
@@ -136,7 +138,7 @@
                 echo "Checking configuration package..."
 
                 cat "${configuration}/config/settings_user.yml" \
-                  | grep -F ${escapeShellArg cfg.stdenv.cc.version}
+                  | grep -F ${escapeShellArg (inputs.conan-flake.lib.pnameFromStdenvCc lib cfg.stdenv)}
 
                 cat "${configuration}/config/profiles/default" \
                   | grep -F "build_type="${escapeShellArg profiles.settings._.build_type}