+++ test/flake-parts-override-default/flake.nix
@@ -60,8 +60,10 @@
             };
 
             settings.compiler = {
-              "${llvmPackages.libcxxStdenv.cc.cc.pname}".__assign = null;
-              "${backendStdenv_13_2.cc.cc.pname}".version.__append = [ backendStdenv_13_2.cc.version ];
+              "${inputs.conan-flake.lib.pnameFromStdenvCc lib llvmPackages.libcxxStdenv}".__assign = null;
+              "${inputs.conan-flake.lib.pnameFromStdenvCc lib backendStdenv_13_2}".version.__append = [
+                (inputs.conan-flake.lib.pnameFromStdenvCc lib backendStdenv_13_2)
+              ];
             };
 
             offline = true;
@@ -107,15 +109,15 @@
                       echo "Checking local setup..."
 
                       cat ${escapeShellArg cfg.configLocal}"/settings_user.yml" \
-                        | grep -F ${escapeShellArg stdenv.cc.version}
+                        | grep -F ${escapeShellArg (inputs.conan-flake.lib.versionFromStdenvCc lib stdenv)}
                       cat ${escapeShellArg cfg.configLocal}"/settings_user.yml" \
-                        | grep -F ${escapeShellArg backendStdenv.cc.version}
+                        | grep -F ${escapeShellArg (inputs.conan-flake.lib.versionFromStdenvCc lib backendStdenv)}
 
                       ! cat ${escapeShellArg cfg.configLocal}"/settings_user.yml" \
-                        | grep -F ${escapeShellArg llvmPackages.libcxxStdenv.cc.version}
+                        | grep -F ${escapeShellArg (inputs.conan-flake.lib.versionFromStdenvCc lib llvmPackages.libcxxStdenv)}
 
                       cat ${escapeShellArg cfg.configLocal}"/settings_user.yml" \
-                        | grep -F ${escapeShellArg backendStdenv_13_2.cc.version}
+                        | grep -F ${escapeShellArg (inputs.conan-flake.lib.versionFromStdenvCc lib backendStdenv_13_2)}
 
                       cat ${escapeShellArg cfg.configLocal}"/profiles/default" \
                         | grep -F "build_type="${escapeShellArg cfg.final.profiles.settings._.build_type}
@@ -153,15 +155,15 @@
                 echo "Checking configuration package..."
 
                 cat "${configuration}/config/settings_user.yml" \
-                  | grep -F ${escapeShellArg stdenv.cc.version}
+                  | grep -F ${escapeShellArg (inputs.conan-flake.lib.versionFromStdenvCc lib stdenv)}
                 cat "${configuration}/config/settings_user.yml" \
-                  | grep -F ${escapeShellArg backendStdenv.cc.version}
+                  | grep -F ${escapeShellArg (inputs.conan-flake.lib.versionFromStdenvCc lib backendStdenv)}
 
                 ! cat "${configuration}/config/settings_user.yml" \
-                  | grep -F ${escapeShellArg llvmPackages.libcxxStdenv.cc.version}
+                  | grep -F ${escapeShellArg (inputs.conan-flake.lib.versionFromStdenvCc lib llvmPackages.libcxxStdenv)}
 
                 cat "${configuration}/config/settings_user.yml" \
-                  | grep -F ${escapeShellArg backendStdenv_13_2.cc.version}
+                  | grep -F ${escapeShellArg (inputs.conan-flake.lib.versionFromStdenvCc lib backendStdenv_13_2)}
 
                 cat "${configuration}/config/profiles/default" \
                   | grep -F "build_type="${escapeShellArg profiles.settings._.build_type}